feat(ui): Indicate a running job
Introduce a new combination of TailwindCSS classes to better indicate
the currently running job of an ORT run. This combination of classes
makes the running job "pulsing" and slightly larger in the traffic
lights component.

Also change the SCHEDULED status colour throughout the UI into a lighter
blue, instead of the earlier yellow, to make room for a future job
status of "FINISHED WITH WARNINGS".

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>