feat(ui): Always enable Analyzer job for an ORT Run
Currently, the Analyzer job needs to always be included in an ORT run,
so make it mandatory in the run creation form.

As there is a pending issue about uploading pre-created Analyzer results
for a run [1], we implement this with minimal changes to the UI, by
always enabling the Analyzer job switch and preventing to disable it.
This prepares for future changes to the run creation form regarding
the upload functionality.

[1]: https://github.com/eclipse-apoapsis/ort-server/issues/193

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>