fix(ui): Identify ORT runs by their index instead of id
ORT runs have a global id and a repository specific index. Use the index
instead of the id when navigating to a run, as the index is used by the
API endpoint to get ORT run details. For report downloads still the
global id has to be used.

This fixes the issue that the links to ORT runs were broken if runs for
multiple repositories exist. Previsouly, this worked by accident in case
only a single repository was used, because then the id and index are
identical. As soon as there ORT runs for more than one repository the
links were broken.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>