docs: Convert README files from AsciiDoc to Markdown
Convert all readme files in the repository from AsciiDoc to Markdown to
use the same lanugage for all of them. Markdown was chosen over AsciiDoc
because it is the more commonly used format in open source projects and
most developers are familiar with it.

While at it, format all of those files to use the one sentence per line
approach to minimize the diffs when the files are edited.

The conversion was done by converting the files using asciidoctor and
pandoc:

  asciidoctor -b docbook5 README.adoc
  pandoc -f docbook -t gfm README.xml -o README.md \
    --base-header-level=2 --wrap=preserve

Afterwards any problems were corrected manually.

Relates to #256.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>