style(ui): Add OpenAPI JSON file to prettier ignore
The OpenAPI JSON file is generated with a Gradle task and stored under
the UI project. The file is quite well formatted already, so add it to
`.prettierignore` to avoid unnecessary Prettier warnings when running
the `format:check` script. Prettier ignore is used, rather than Git
ignore, for the reasons explained in [1].

[1]: https://github.com/prettier/prettier/issues/15438#issuecomment-1731842806

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>