feat(ui): Change the appearance of the delete buttons and dialogs
In an earlier team review, it was suggested that instead of making the
delete buttons red, they should be made more neutral, and instead render
the opening delete confirmation dialogs with a more warning-like
appearance. Moreover, it was requested to also show the name of the item
which is about to be deleted in that dialog.

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>