feat(ui): Create dummy rerun buttons for each ORT Run
In the repository index page, which lists all the ORT runs for a
repository, add an Actions column and add a dummy rerun button for each
ORT run. The button currently leads to a normal "create run" page;
functionality to pass on job configs to the new run will be added in a
later commit.

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>