feat(Analyzer): Use Corepack to install Yarn2+ and PNPM
Corepack [1] provides an alternative way to install the package
managers it supports (Yarn2+ and PNPM). Projects can specify the
concrete version number of the package manager they need; Corepack
then handles the automatic installation of this version.

For this to work, the package managers must not be installed by any
other means. So, remove them from the container image and explicitly
enable Corepack.

[1]: https://nodejs.org/api/corepack.html

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>