feat(ui): Add an indeterminate state to the checkbox component
All checkbox lists in the UI have also an "enable/disable all" checkbox
on top of the list. In team review it was noted that the component also
should indicate when only some of the checkboxes are checked.

By default, the shadcn-ui checkbox component [1] doesn't support this,
but it is based on Radix UI [2], which has "true/indeterminate/false"
state support for checkboxes. shadcn-ui components can be freely
modified, so add to the checkbox component support for the indeterminate
state using the "group" modifier from TailwindCSS. Also add handling of
the indeterminate state to all checkbox lists in the "create ORT run"
form.

[1]: https://ui.shadcn.com/docs/components/checkbox
[2]: https://www.radix-ui.com/primitives/docs/components/checkbox#api-reference

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>