feat(config): Add a git config file provider
Add an implementation of `ConfigFileProvider` that reads the config files
from a Git repository and stores them to a local directory.

All unit tests are run using a synthetic test repository
https://github.com/doubleopen-project/ort-config-test. The structure of
the repository allows for testing access to files, folders, symlinks and
submodules.

Also add the needed runtime dependencies and configuration properties
to all workers.

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>