feat(core): Set Swagger UI server URL from config
This change introduces three new configuration variables that are used
for setting the "server" value in the Swagger UI configuration. The new
variables have default values that keep the existing behaviour.

This enables using Swagger UI in environments with some kind of proxy
or load balancer in front of the ORT Server API.

Signed-off-by: Haiko Schol <hs@haikoschol.com>