feat(workers): Add functionality to generate proxy configuration
Multiple package managers support proxy settings in their
configuration files, but so far, only the generator for the NPM
configuration can handle this. To simplify the implementation for
other generators, provide corresponding functionality in the
`ConfigFileBuilder` class. Add a function here that checks whether a
proxy for the system is defined, and - if so - execute a block to
generate a proxy configuration.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>