fix(core): Fix a race condition when creating ORT runs
Creating multiple runs for the same repository concurrently could lead
to a unique constraint violation exception because the computation of
the run index is not thread-safe. Fix this by setting the transaction
isolation level to `TRANSACTION_SERIALIZABLE`. Since runs are not
created that frequently, this should not have a major impact on the
performance.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>