refactor(workers): Rename extension functions
Rename the `readConfigFile` and `readConfigFileWithDefault` extension
functions to `readConfigFileValue` and `readConfigFileValueWithDefault`
to make it clearer that they not only read the files but also
deserialize the content, and to make room for similar helper functions
that do not perform deserialization.

The new function names are inspired by the `readValue` function from the
Jackson Kotlin module.

While at it, also slightly improve the function docs.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>