fix(scanner): Fix assignments of provenances to nested provenances
OrtServerNestedProvenanceStorage did not handle the constellation
correctly that there are multiple root projects in a repository. In
this case, only the last project was assigned the nested provenance
correctly; the others were missed.

To prevent this, extend `PackageProvenanceCache` to store a list of
provenance IDs that are found in a repository. Then the assignment can
be done for all of these provenances.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>