fix(ui): Redirect back to same path after signing in
To sign the users in, the UI briefly redirects the browser to Keycloak when
loading the UI in browser. Keycloak redirects users back to the UI, so use
the current URL as the redirect URI to get the user to the same page they
tried to access before signing in.

Fixes #435.

Signed-off-by: Mikko Murto <mikko.murto@doubleopen.org>