fix(job-monitor): Handle Notifier jobs when searching for lost jobs
When the Notifier worker was added, it was missed to update the Job
Monitor component to include the new worker in the lost job detection
mechanism. Add it now and extend `LostJobsFinderTest` to no longer
compile when a new endpoint is added that might need to be handled; so
that this mistake is not repeated in the future.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>