feat(config): Deal with GitHub rate limits
In `GitHubConfigFileProvider`, check whether a request failed because
rate limits [1] were exceeded. If so, retry the request after the next
reset of the rate limits.

Note that this solution only helps if rate limits are reached from
time to time at the end of a rate limit interval. Otherwise, waiting
for the next reset could be rather long.

[1]: https://docs.github.com/en/rest/using-the-rest-api/rate-limits-for-the-rest-api?apiVersion=2022-11-28

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>