feat(config): Add a GitHubConfigCache interface
The purpose of this interface is to enable caching for configuration
data downloaded from a GitHub repository. It is going to be used by
`GitHubConfigFileProvider` to reduce its dependency to the GitHub API.

An interface is used here, as there are multiple planned
implementations: A functional one based on a file-cache and one that
basically disables caching.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>