feat(ui): Add parameters and labels to run creation form
Allow to specify parameters and labels in the run creation form. They
need to be added as key-value pairs of strings, which is also validated
in the form with error messages to the user in case of validation
failures. As it is unclear at the moment, whether values can be empty or
not, we allow empty values for now.

Only add and delete methods need to be specifically added as buttons,
as editing of existing parameters and labels can be done inline.

If the user enters this page by pushing a Rerun button for an earlier
run, the parameters and labels are copied from it.

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>