test(core): Fix race condition when creating ORT runs again
Increase the number of concurrent requests to 50 in the test for
creating ORT runs concurrently. This shows that the fix in d517f0d was
not sufficient, as the following PostgreSQL error can still occur:

  could not serialize access due to read/write dependencies among
  transactions

To fix the issue, increase the max attempts for the transaction from the
default value 3 to 20. This does not completely prevent the error to
occur but makes it a lot less likely. Also, in practice there should not
be a reason to create that many ORT runs on the same repository at the
same time.

20 repetition attempts might seem a lot, but creating an ORT run is an
operation that should rather take a bit longer than to fail.

For reference, see the PostgreSQL documentation which says that
applications using the serializable isolation level "must be prepared to
retry transactions due to serialization failures" [1].

[1]: https://www.postgresql.org/docs/current/transaction-iso.html#XACT-SERIALIZABLE

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>