feat(ui): Add some configuration options to the evaluator job
Add the possibility to specify evaluator rules, license
classifications, copyright garbage, and resolutions files to the
evaluator job. The user can also specify a path (relative to the root
directory of the config repository) where the configuration file is
located.

In case the path&file is not specified, the corresponding config
parameter structure is not passed to the evaluator and it will look for
the configuration file in the default directory of the config provider.

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>