fix(config): Address a race condition in `GitHubConfigFileCache`
There is a certain flakiness in the tests of the file cache which
indicates a potential problem. Obviously it can happen that a process
that wants to read a file from the cache is able to obtain a read lock
before the downloading process gets a write lock. Then an empty file
is read. Try to fix this by checking the file length and retrying the
operation if necessary.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>