refactor(vault): Remove an unneeded `runBlocking` call
Rework the token handling to not fetch a token initially, but deal
with a missing token during request execution. While this makes the
logic slightly more complex, it has the benefit that during creation
of a `VaultSecretsProvider` instance no request is sent, and the
`runBlocking` call to fetch the initial token can be dropped.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>