feat(api): Add a `PagingData` class for responses
Add a separate `PagingData` class which holds the paging data used for
generating the response. The class will replace `PagingOptions` for
responses in later commits.

Having a dedicated response class has the benefit that the properties
do not have to be nullable which simplifies handling the response. Also,
the class has a property for the total count of entities which would not
make sense for requests.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>