feat(api): Temporarily add a `PagedResponse2` class
Add a new data class for paged responses that uses the `PagingInfo`
class introduced in the previous commit. This allows to iteratively
migrate to the new response type without having to apply the changes for
all endpoints in a single commit. After the migration, the class will
replace the existing `PagedResponse` class.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>