feat(api): Migrate `/organizations/{id}/products` to `PagedResponse2`
Migrate the endpoint to get the products for an organization to
`PagedResponse2` to provide the total entity count. To implement this,
add a new `ListQueryResult` class to return the total count also as part
of the database query.

Getting the total count is implemented as a separate database query
which should be very fast as the `organization_id` column is indexed.
While it would be possible to implement getting the total count using a
window function [1], this does not necessarily provide better
performance and increases the size of the result set returned from the
database [2].

[1]: https://github.com/JetBrains/Exposed/issues/1576#issuecomment-1866761186
[2]: https://stackoverflow.com/a/28888696/639421

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>