refactor(ui): Add a component for multiselection from a list of options
Add a reusable component to handle multiselection lists in forms.
Checkboxes are used for the selection. The component also handles
enabling and disabling of all selections with one click, as well as
showing "intermediate" state for the enable/disable element when some
(but not all) of the items in the list are selected.

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>