fix(ui): Fix a bug in Notifier job when using rerun functionality
The server back-end stores recipient addresses in a string array, while
the ORT run creation form uses a useFieldArray hook [1], which deals
with arrays of objects. When using rerun functionality, the recipient
addresses of the earlier run were not correctly populated in the form
but instead always initialized to an empty array.

[1]: https://react-hook-form.com/docs/usefieldarray

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>