fix(ui): Fix uncontrolled/controlled errors in the ORT Run creation form
When text input fields in the form are left uninitialized, React throws
a warning about a component entering from uncontrolled to controlled
state. This has no effect on the functionality of the form, but the
warning is seen in the browser's development console.

This commit fixes the issue by initializing optional input fields in
the form with some default (mostly empty) values. If the user leaves an
optional field empty when submitting the form, an `undefined` is sent,
in effect leaving the configuration parameter out of the request.

[1]: https://stackoverflow.com/questions/77201810/shadcn-input-form-zod-a-component-is-changing-an-uncontrolled-input-to-be-c

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>