fix(ui): Wrong base default value given in Notifier config
PR #670 introduced a bug in the Notifier configuration, where the base
default value of Notifier's recipient addresses was set to an empty
object array with empty objects.

This actually caused the Notifier subform to complain about a missing
e-mail address, but the message was hard to see, because Notifier is
disabled by default, so the user would expect not to touch Notifier job
configuration, and the corresponding accordion is closed in the
UI. In effect, all create ORT run actions failed, due to the (closed)
subform not passing validation. By fixing the default value to an empty
array, the Notifier subform will not complain about missing e-mail.

Fixes #681.

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>