feat(ui): Add new run details page
To isolate the new page from existing implementation, use a temporary
route `.../repositories/{repoId}/runs2` for new implementation.

Layout (pathless) route is used to define a layout that is shared by
all its child routes. This is used to include a sidebar that is present
in run details pages as well as their subroutes.

At this stage the sidebar acts as a placeholder only, and will be
populated in a later PR.

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>