fix(ui): Guard against possible undefined context in header
For some reason, the context becomes undefined when the browser is
refreshed on organization, product, repository and run index pages,
throwing an error. TypeScript didn't catch this, and breadcrumb
functionality isn't affected as such, so the error only appears runtime
when refreshing.

The underlying reason could be somewhere deeper, but for now, adding a
check for a defined context prevents the error from happening.

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>