feat(ui): Implement a dashboard-like view for the run details page
Other pages in the ORT Server UI are allowed to freely spill over
vertically with more elements in the page, as scrolling action for them
is natural. However, the run details page is a special case, as it is
of dashboard nature so should always fit into the viewport of the
browser without spilling over and causing scrolling.

Add a special calculation of the page height to the layout route of the
new run details pages, which accommodates for the header height and
padding given in the `header` and `main` classes. Also add overflow
guarding and some more `lorem` content to the Issues component, to
demonstrate content scrolling.

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>