feat(ui): Add generic DeleteDialog component
The delete buttons and the dialogs that are opened by clicking them
in the current UI all follow a similar pattern, so a generic
component is created here, which can be used to reduce some code
clutter.

The description part of the dialog could be altered to take in a JSX
element if needed, but as all current dialogs use the same pattern,
only take in the changing parts as props for now.

The open state of the dialog is handled with a React state variable,
to better be able to handle the open state based on API call states
and responses for example.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>