fix(ui): Fix active link not highlighting correctly
The active path link was not highlighted properly in the sidebar if
there were search params in the URL. Add includeSearch parameter
with value 'false' to the activeOptions of the Link component to fix
this issue. While at it, make the activeOptions configurable to
allow for different behaviour on different use cases of the sidebar.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>