fix(reporter): Handle sub paths in the source code bundle reporter
`SourceCodeBundleReporter` used to ignore the `path` attribute in the
VCS info of packages downloaded from a source code repository.
Therefore, the source code archives for such packages included the
whole content of the repository. Fix this by checking for a path, and
only including files that are direct or indirect children of this
path.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>