refactor(transport): Use `suspend` functions as entry points
Change `MessageReceiverFactory` - which is the starting point for all
message processing in ORT Server components - to use `suspend`
functions. That way, the processing logic can directly deal with other
coroutine-friendly APIs without any quirks like `runBlocking` calls.

The goal is to rework more and more APIs of ORT Server components and
abstractions to be based on coroutines and `suspend` functions, so
that this programming model can be used everywhere in the codebase.

This change also impacts tests. To fix issues here, convert the test
helper functions `mockTransaction` and `withMockDatabaseModule` to
`inline` functions. This is the easiest (and given their size probably
also the most efficient) way to support invocations of `suspend`
functions in the blocks they execute.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>