refactor(job-monitor): Introduce a `TimeHelper` class
The class contains common logic to query the current system time and
to compute timestamps in the past that are used to filter for jobs of
a specific age. This functionality is going to be used by another
component as well; therefore, it is helpful to have it available in a
helper class.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>