deps: Downgrade Vite to 5.3.5
With Vite 5.4.0 the `VITE_` configuration variables do not appear
anymore in the built JS file. For example, instead of:

    Pe.BASE=nd.VITE_API_URL||"http://localhost:8080";

the output file now contains only:

    Pe.BASE="http://localhost:8080";

This breaks the entrypoint script for the UI Docker image which makes it
impossible to use the image with other than the default values.

Downgrade Vite to 5.3.5 until a proper solution is found.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>