fix(ui): Fix query keys
Some query keys include functions instead of string variables, which
leads to nulls being used as keys. This causes some queries sharing the
same cache, leading to inconsistent data being shown sometimes in the
UI. Fix this by always importing and using from API queries a variable
ending with Key, which is a string.

As described in [1], the string array that is used as the query key
should always include the variables used in the query function, so also
fix this in all the affected files.

Resolves #812.

[1]: https://rb.gy/3m08yc

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>