fix(core): Exclude `OptionalValue` from the OpenAPI spec
The `OptionalValue` class is only used in Kotlin code to model the
difference between `null` values and not present values for PATCH
requests. In the JSON request bodies such properties can simply be
omitted.

Configure type redirects for `OptionalValue` to its generic types used
in the API classes to eliminate the class from the generated OpenAPI
spec.

Fixes #821.
Relates to #605.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>