fix(dao): De-duplicate packages when creating an Analyzer run
It is intended that multiple Analyzer runs detecting the same package
should use a shared reference to the same database entity. Due to a
bug in `PackagesTable`, this deduplication did not work, however.
Because of this problem, lots of redundant packages were created, and
the packages table grew enormously.

To prevent this, fix the deduplication logic in `PackagesTable` and
related problems when inserting a new package. Add a helper function
for comparing child entities, as this was the source of the problem.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>