fix(dao): De-duplicate projects when creating an Analyzer run
`ProjectsTable` was suffering from the same problem as `PackagesTable`
related to the deduplication of entities. In addition, the table only
had a 1:n relation to the Analyzer runs. In order to support sharing
of projects entities, replace this by an m:n relation. Add a
corresponding migration script.

The migration script contains another fix that is required for older
databases only. Here, there could be some inconsistencies in the
relations to processed licenses which cause errors with the now fixed
de-duplication algorithm.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>