fix(dao): Do not load unnecessary data when creating ORT run
The code to determine the next run index when creating a new ORT run was
unnecessarily loading the complete list of ORT runs for the repository
and mapping it to the business model. Therefore a lot of unnecessary
data was loaded not only from the `ort_runs` table, but also from
associated tables like `labels` or `issues`.

To fix this, replace the call to `listForRepository` with a custom query
that loads only the last index. This increases performance especially
for repositories with a lot of runs and makes it also a lot less likely
that the transaction has to be reattempted when runs are created
concurrently.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>