feat(ui): Add a new toast component
For better control over the info and error toasts in the UI, switch to
using the Sonner component [1] from shadcn-ui instead of the usual Toast
component.

To make sure the user catches the errors from back-end, for
example when invalid data is being sent in the requests, make the error
toasts have a specific dismiss button so the error doesn't disappear
before the user has pressed the button in the toast.

[1]: https://ui.shadcn.com/docs/components/sonner

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>