fix(dao): Make deduplication logic more robust
In the find functions that are used in the `getOrPut` functions to
obtain references to deduplicated entities, replace `singleOrNull` by
`firstOrNull`. If the deduplication works as expected, this change
does not make any difference. If there are duplicate entries, however,
`firstOrNull` is more likely to prevent the creation of new
duplicates. Also, from the semantics of find functions, it is rather
strange to return `null` although there are matching entries.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>