fix(ui): Fix applying the config variables in the entrypoint script
Since Vite 5.4.0 the `VITE_` variables are evaluated during the build,
therefore the replacements used in the entrypoint script do not work
anymore.

To fix this, set placeholder values for those variables when building
the UI for the Docker image and adapt the entrypoint script to replace
those placeholder values.

While at it, also sort the variable handling in the entrypoint script
alphabetically for better overview.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>