feat(ui): Download logs from an ORT Run
Add a download button for the logs of an ORT run. Note that this does
not currently work in Docker Compose environment, because of the missing
`runId`, so all logs in the downloaded archive are empty.

As a first quick implementation, no "level" and "steps" are passed to
the query, so default values are used instead:
  level = INFO
  steps = all steps of the run

This is mainly to get something quickly to the deployment environment,
to test the feature out. If the log downloading works, subsequent
PRs will fine-tune this feature.

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>