test(transport): Increase the queue timeout
In CI, `SqsMessageReceiverFactoryTest` often fails with "Expected value to
not be null, but was null" when polling the queue [1], which does return
"null if the specified waiting time elapses before an element is
available".

Triple the timeout to 15 seconds to try addressing that, and move the
timeout to a constant while at it.

Hopefully fixes #720.

[1]: https://github.com/eclipse-apoapsis/ort-server/actions/runs/10488199036/job/29050829407?pr=861#step:5:3552

Signed-off-by: Sebastian Schuberth <sebastian@doubleopen.org>