chore(workers): Remove unnecessary calls to `runBlocking`
After the `EndpointHandler` became a suspend function in 7bab95c some
`runBlocking` calls have become unnecessary and it became trivial to
turn some functions into suspend functions to remove most of the
remaining `runBlocking` calls.

Removing calls to `runBlocking` is important, because the function does
not preserve the MDC context.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>