fix(workers): Use ORT's version of `runBlocking`
Replace the remaining `runBlocking` calls in the workers with calls to
ORT's version of `runBlocking` which preserves the MDC context.

Note that the server uses the SLF4J API to access the MDC while ORT uses
the Log4j API. This still works because with `log4j-to-slf4j` on the
classpath, Log4j uses SLF4J's MDC implementation.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>