fix(dao): Add a migration to drop duplicate packages and projects
Due to an issue fixed in fb57726, duplicates of packages and projects
were stored in the database. This consumes unnecessary space and makes
it hard to gather statistics. Also, the new mechanism to prevent
duplicates is very slow if the database contains a lot of packages
(>1,000,000).

The script is rather complex, and its execution can take a while on a
large database. This might delay the start of the server when the
migration has to run.

So add a migration script to remove the duplicates.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>