feat(ui): Use a new component to show duration of ongoing runs
Create a `RunDuration` component to also show the current duration of
an ongoing ORT run. An ongoing run is distinquished visually from the
completed runs with a pulsing italic duration string.

The updating interval of the run duration can be configured, and
currently, it is set to the `VITE_RUN_POLL_INTERVAL` value (defaulting
to 10 seconds).

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>