feat(core): Add an API for commands and queries
Add base classes for commands and queries and the respective handlers.
This is a preparation for testing if CQS [1] would improve the structure
of the code, mainly with regards to separation of concerns and
testability.

Another problem that this approach is supposed to solve is that queries
often combine data from different aggregate roots. As each `Repository`
is responsible to ensure data consistency within the aggregate root it
represents, combining data is currently often done via different
repositories which is inefficient compared to writing a custom query.

When code that changes data and code that only reads data is clearly
separated in commands and queries, it would be easier to establish a
convention where all commands have to use repositories to modify data,
but queries are free to use custom SQL.

[1]: https://en.wikipedia.org/wiki/Command%E2%80%93query_separation

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>