feat(ui): Clicking traffic lights takes to corr. job configuration
Add section anchors to the job configuration sub-page of the ORT run
details page. Clicking a traffic light of a job in the repository
index page takes the user to the corresponding section of the job
configuration page.

NOTE: The id's that mark the position in the scrolling job configuration
page need a `scroll-mt-16` class to compensate for the UI's header,
otherwise clicking a job traffic light would result to the page opening
such that the job configuration is partly scrolled behind the header.

Resolves #643.

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>