feat(ui): Show labels and parameters in a more readable way
Improve readability of ORT run results by showing the labels and
parameters each in their own row, using a grid layout that aligns the
rows on "=".

Also save some screen real estate - in anticipation of more cards coming
to the results main page - by using an extra small font for the labels
and parameters.

If needed, the layout could be improved in future for example by
finding the longest string from labels and parameters of a run and using
this to set the width of the (leftmost) key column.

Resolves #937.

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>