feat(ui): Create an issues page for ORT run
Also add a sidebar link to the page. Please note that this page is by no
means complete yet, as it's just using the `ìssues` object of an ORT
run, which only includes issues created by the ORT Server. The page
should be viewed as a bookmark for future development.

When issue #405 is being worked on, the page should be updated to use
the issues endpoint of the ORT Server API.

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>