test(transport): Increase the timeout when querying the blocking queue
In the ORT Server CI, the `SqsMessageReceiverFactoryTest` is flaky with
an error that no message is present in the blocking queue.

Since the test takes some time to initialize with the creation of a new
thread to receive messages from the transport, increase the timeout to
try to alleviate this behavior.

Signed-off-by: Nicolas Nobelis <nicolas.nobelis@bosch.com>