feat(core): Add admin endpoint to trigger Keycloak sync
Add a new endpoint /api/(v1/admin/sync-roles to manually trigger
the synchronization of Keycloak roals and groups.

This is useful when Keycloak got out of sync with the server,
for example, because there were temporary connection issues.
Previously, it was required to restart the server to
trigger a new sync.

Signed-off-by: Kamil Bielecki <kamil.bielecki@pl.bosch.com>