fix(core): Use valid references in the OpenAPI spec
According to the specification [1], keys used in the schemas map must
match the regular expression `^[a-zA-Z0-9\.\-_]+$`. By default,
ktor-schema-kenerator-swagger generates the names for schemas of generic
types with angle brackets and comma-separated type arguments, for
example, `GenericType<Type1,Type2>`.

As the generation of those names is currently not configurable [2], copy
and adapt some classes of the library to use only valid characters, for
example, `GenericType_Type1-Type2`.

As long as the copied functions are required, add a copyright statement
to the file header. If the problem gets fixed upstream, this can be
removed again.

Resolves #605.

[1]: https://swagger.io/specification/#components-object
[2]: https://github.com/SMILEY4/schema-kenerator/issues/13

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>