feat(ui): Use custom hooks for queries
OpenAPI React Query Codegen [1] provides custom hooks that we have been
using for the PATCH queries. Switch to using custom hooks also for GET
queries. These custom hooks wrap the query keys and query functions
conveniently inside the hook, avoiding repetitive boilerplate code and
making code generally easier to read.

For route loaders, switch to using the prefetch variety of hooks.
Prefetching in general is a mechanism for loading data needed for a
route's component tree ahead of time. Using prefetching queries in route
loaders leaves us freedom in deciding for each route, which prefetch
queries are blocking and which not [2], thus optimizing the user
experience.

A good example for optimising the organization route would be to do a
blocking prefetch for organization data (name, description), to quickly
show the layout and main data for the user, followed by a non-blocking
prefetch for the product list.

Note that the logic of prefetching data for routes has not been tuned at
this point, but all prefetching is still done in a blocking manner.
This can easily be tuned later on, when we gather more user experience
from real-world server deployments.

[1]: https://github.com/7nohe/openapi-react-query-codegen
[2]: https://rb.gy/233olf

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>