feat(ui): Add helpers for presenting overall vulnerability status
Add some helper functions, types and constants to help calculating and
presenting the overall (qualitative) rating of a vulnerability [1].

Please note that the first implementation of the overall rating
calculation mimics the simple calculation in the ORT WebApp report, and
as such should at this stage be regarded as advisory only. The main
purpose of it is to provide some visual guidance to the viewer.

For future implementation, using a dedicated CVSS calculation library is
under consideration, such as `ae-cvss-calculator` [2]. Preparing for
this, TailWindCSS color classes to represent the different vulnerability
ratings are chosen to match the library as closely as possible.

The chosen colors are based on the TailWindCSS background color palette
[3], and are as follows:
  NONE:     bg-neutral-300
  LOW:      bg-yellow-400
  MEDIUM:   bg-amber-500
  HIGH:     bg-orange-600
  CRITICAL: bg-red-600

[1]: https://www.balbix.com/insights/understanding-cvss-scores/
[2]: https://www.npmjs.com/package/ae-cvss-calculator
[3]: https://tailwindcss.com/docs/background-color

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>